refactor: :recycle: update atomic components for dashboard
