+++ src/dashboard/presenters/components/index.ts
@@ -1,2 +1,3 @@
+export * from "./molecules";
 export * from "./organisms";
-export * from "./template";
+export * from "./templates";
