+++ src/dashboard/presenters/components/organisms/ListPeople.tsx
@@ -1,10 +1,12 @@
 import { FindPeople } from "@/dashboard/data";
 
-import { PersonCard } from "../organisms/PersonCard";
+import { PersonCard } from "../molecules/PersonCard";
 import { personRepositoryFactory } from "../../factories";
 
+	// This would be an organism because it has some molecules
 export async function ListPeople() {
 	const peopleFound = await new FindPeople(personRepositoryFactory()).execute();
+
 	return (
 		<div className="flex flex-wrap w-320 gap-6">
 			{peopleFound.map(card => (
