--- src/dashboard/presenters/components/template/index.ts
@@ -1,1 +1,0 @@
-export * from "./ListPeople";
